+++ postcss.config.js
@@ -1,11 +1,6 @@
-// Copyright 2011-2019 The Bootstrap Authors
-// Copyright 2011-2019 Twitter, Inc.
-// Licensed under MIT (https://github.com/twbs/bootstrap/blob/master/LICENSE)
-// Source: https://github.com/twbs/bootstrap/blob/v4-dev/build/postcss.config.js
-
 'use strict'
 
-module.exports = (ctx) => ({
+module.exports = {
   map: {
     inline: false,
     annotation: true,
@@ -16,4 +11,4 @@ module.exports = (ctx) => ({
       cascade: false
     }
   }
-})
+}