+++ CHANGELOG.rst
@@ -19,6 +19,7 @@ Latest
 * Changed SCSS entry file name from ``theme.scss`` to ``main.scss``.
 * Added ``include-path node_modules/`` option to node-sass command line. This
   allows more concise imports of Bootstrap sources in our SCSS files.
+* Cleaned up the PostCSS config file.
 
 
 Version 1.1